refactor(graph): extract validation and input_spec into graph/ package
- Move Graph class from graph.py to graph/core.py (~300 lines)
- Extract validation logic to graph/validation.py (~250 lines)
- Extract InputSpec and calculator to graph/input_spec.py (~150 lines)
- Add graph/__init__.py for clean re-exports

Graph class reduced from ~600 lines to ~300 lines by extracting:
- All validation methods (9 functions) to dedicated module
- Input specification computation (8 functions) to dedicated module

Public API unchanged - users still import from hypergraph directly.
