+++ src/hypergraph/graph/__init__.py
@@ -1,0 +1,13 @@
+"""Graph package - core graph structure and validation."""
+
+from hypergraph.graph.core import Graph
+from hypergraph.graph.validation import GraphConfigError, validate_graph
+from hypergraph.graph.input_spec import InputSpec, compute_input_spec
+
+__all__ = [
+    "Graph",
+    "GraphConfigError",
+    "InputSpec",
+    "compute_input_spec",
+    "validate_graph",
+]
